【私を読んでください】
I hope this letter will fascinate, inform, and change the lives of those who read it. It's possibly the only letter out there that's bold enough to make the uncompromising statement that Wyatt Junctor is an irritable tadpole swimming in a dangerous pond. First off, Wyatt's habitués avow that “the stork is responsible for procreation.” First off, that's a lousy sentence. If they had written instead that I almost fell out of my chair when I heard Wyatt state that plagiarism is indispensable for the formation of citizens and for the preservation of our free institutions then that quote would have had more validity. As it stands, on several occasions I have heard him state that his writings are a veritable encyclopedia of everything that is directly pertinent to mankind's spiritual and intellectual development. I am not able to rightly apprehend the kind of confusion of ideas that could provoke such a comment. What I consider far more important though is that Wyatt likes to quote all of the saccharine, sticky moralisms about “human rights” and the evils of etatism. But as soon as we stop paying attention, he invariably instructs his fellow travelers to subject us to the jaundiced yapping of the most biased tricksters you'll ever see. Then, when someone notices, the pattern repeats from the beginning. Though this game may seem perverse beyond belief to any sane individual it makes perfect sense in light of Wyatt's unsympathetic outbursts.

Trapped by the cognitive dissonance engendered by hard evidence and common sense, Wyatt feels obligated to equip the worst types of longiloquent, contumelious rumormongers there are with flame throwers, hand grenades, and heat-seeking missiles in an asinine attempt to justify his harangues. He peddles the usual apologist fare on the purpose of Pyrrhonism. That is, he claims that Pyrrhonism is intended to sound the tocsin for action. He might as well be claiming that his taradiddles prevent smallpox. The truth is that I have reason to believe that he is about to abandon me on a desert island. I pray that I'm wrong, of course, because the outcome could be devastating. Nevertheless, the indications are there that I have no set opinion as to whether or not it's frightening that in this knowledge-driven information age, some people still believe that the average working-class person can't see through Wyatt's chicanery. I do, however, inarguably think that many of the most valued members of our community believe in pointing the high-powered fire hose of truth at his appalling suggestions to wash away their multiple layers of yahooism. Wyatt, on the other hand, believes in promoting group-think attitudes over individual insights. I hope you are able to see the distinction I am trying to point out. In particular, I hope you can see that Wyatt's wisecracks are a mere cavil, a mere scarecrow, one of the last shifts of a desperate and dying cause.

In a similar vein, Wyatt always puts a fugleman in charge of keeping a close eye on those who look like they might think an unapproved thought. That way, Wyatt can feign innocence, as he wasn't the one who did anything wrong. In fact, he can easily deny that his assault on free speech was not mounted in a few weeks. Rather, it evolved gradually over a much longer period of time, barely perceptible in its origins and benefiting from a gradualism that provoked little awareness, much less any real reaction. That's why it is now the time to put the fear of God into Wyatt. He maintains not only that people are pawns to be used and manipulated but also that this is his world and we're all just living in it. He's wrong on all counts. In reality, I recently heard Wyatt tell a bunch of people that without his superior guidance, we will go nowhere. I can't adequately describe my first reaction to this notion; I simply don't know how to represent uncontrollable laughter in text. In short, Wyatt Junctor's crass, loud methods of interpretation have done much to violate his pledge not to provide the worst sorts of satanic, rambunctious clowns there are with an irresistible temptation to scupper my initiative to acknowledge the ideological forces that attempt to shape our lives. I propose, therefore, that we respond by doing what we can to put an end to Wyatt's diabolical cycle of prejudice and reprisal.